Changing as per review comments.

+++ src/app/util/af-types.h
@@ -243,7 +243,7 @@ struct EmberAfDefinedEndpoint
      */
     char endpointUniqueId[chip::app::Clusters::Descriptor::Attributes::EndpointUniqueID::TypeInfo::MaxLength()] = { 0 };
 
-    size_t endpointUniqueIdSize = 0;
+    uint8_t endpointUniqueIdSize = 0;
 #endif
 };
